+++ src/Braintree/Test/Nonce.cs
@@ -16,7 +16,6 @@ namespace Braintree.Test
         public const string AndroidPayAmEx = "fake-android-pay-amex-nonce";
         public const string AmexExpressCheckout = "fake-amex-express-checkout-nonce";
         public const string AbstractTransactable = "fake-abstract-transactable-nonce";
-        public const string Europe = "fake-europe-bank-account-nonce";
         public const string Coinbase = "fake-coinbase-nonce";
         public const string VenmoAccount = "fake-venmo-account-nonce";
         public const string TransactableVisa = "fake-valid-visa-nonce";